TS-228: partial checkin, fixing more uses of "long". Also removed a dead file which usesd "long"
git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1037833 13f79535-47bb-0310-9956-ffa450edef68

+++ libinktomi++/ink_string++.cc
@@ -214,7 +214,7 @@ StrList::_new_cell(const char *s, int len_not_counting_nul)
     p = (char *) alloc(sizeof(Str) + 7);
     if (p == NULL)
       return (NULL);            // FIX: scale heap
-    p = (char *) (((long) (p + 7)) & ~7);       // round up to multiple of 8
+    p = (char *) ((((uintptr_t)p) + 7) & ~7);       // round up to multiple of 8
     cell = (Str *) p;
   }
   ++cells_allocated;
